0x11. C - printf README commit

+++ _printf_num.c
@@ -8,8 +8,8 @@
  */
 unsigned int printf_di(va_list val, int counter)
 {
-	int num = va_arg(val, int);
-	unsigned int revNum;
+	int long num = va_arg(val, int);
+	unsigned long int revNum;
 	int i = 0, j = 0;
 	char table[41];
